+++ src/main/java/com/editora/ui/SettingsWindow.java
@@ -218,6 +218,7 @@ public class SettingsWindow {
     private CheckBox csvGridCheck;
     private CheckBox structuredPreviewCheck;
     private CheckBox svgPreviewCheck;
+    private CheckBox crontabPreviewCheck;
     private CheckBox csvRainbowCheck;
     private CheckBox autoRenameTagCheck;
     private CheckBox autoCloseTagsCheck;
@@ -929,6 +930,7 @@ public class SettingsWindow {
         csvGridCheck = viewCheck(tr("settings.csvPreview"), Settings::setCsvPreview);
         structuredPreviewCheck = viewCheck(tr("settings.structuredPreview"), Settings::setStructuredPreview);
         svgPreviewCheck = viewCheck(tr("settings.svgPreview"), Settings::setSvgPreview);
+        crontabPreviewCheck = viewCheck(tr("settings.crontabPreview"), Settings::setCrontabPreview);
         csvRainbowCheck = viewCheck(tr("settings.csvRainbow"), Settings::setCsvRainbow);
         autoRenameTagCheck = viewCheck(tr("settings.autoRenameTag"), Settings::setAutoRenameTag);
         autoCloseTagsCheck = viewCheck(tr("settings.autoCloseTags"), Settings::setAutoCloseTags);
@@ -2008,6 +2010,8 @@ public class SettingsWindow {
                 "json yaml toml openapi swagger tree structured data preview");
         Label svg = section(p, tr("settings.section.svg"));
         row(p, Category.EDITOR, svg, svgPreviewCheck, "svg image vector graphic preview render");
+        Label crontab = section(p, tr("settings.section.crontab"));
+        row(p, Category.EDITOR, crontab, crontabPreviewCheck, "crontab cron schedule preview next run times");
         Label saving = section(p, tr("settings.section.saving"));
         Label delayLabel = note("delay (seconds)");
         HBox autoSaveBox = new HBox(8, autoSaveCombo, autoSaveDelaySpinner, delayLabel);
@@ -5414,6 +5418,7 @@ public class SettingsWindow {
             csvGridCheck.setSelected(settings.isCsvPreview());
             structuredPreviewCheck.setSelected(settings.isStructuredPreview());
             svgPreviewCheck.setSelected(settings.isSvgPreview());
+            crontabPreviewCheck.setSelected(settings.isCrontabPreview());
             csvRainbowCheck.setSelected(settings.isCsvRainbow());
             autoRenameTagCheck.setSelected(settings.isAutoRenameTag());
             autoCloseTagsCheck.setSelected(settings.isAutoCloseTags());
@@ -5990,6 +5995,7 @@ public class SettingsWindow {
             csvGridCheck.setSelected(s.isCsvPreview());
             structuredPreviewCheck.setSelected(s.isStructuredPreview());
             svgPreviewCheck.setSelected(s.isSvgPreview());
+            crontabPreviewCheck.setSelected(s.isCrontabPreview());
             csvRainbowCheck.setSelected(s.isCsvRainbow());
             autoRenameTagCheck.setSelected(s.isAutoRenameTag());
             autoCloseTagsCheck.setSelected(s.isAutoCloseTags());